+++ src/Repository/Evaluation/ResourceRepository.php
@@ -7,10 +7,10 @@ use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
 
 /**
- * @method Resource|null find($id, $lockMode = null, $lockVersion = null)
- * @method Resource|null findOneBy(array $criteria, array $orderBy = null)
- * @method Resource[]    findAll()
- * @method Resource[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @method resource|null find($id, $lockMode = null, $lockVersion = null)
+ * @method resource|null findOneBy(array $criteria, array $orderBy = null)
+ * @method resource[]    findAll()
+ * @method resource[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
 class ResourceRepository extends ServiceEntityRepository
 {