+++ src/Security/Voter/PlaceVoter.php
@@ -26,7 +26,7 @@ class PlaceVoter extends Voter
 
     protected function voteOnAttribute($attribute, $place, TokenInterface $token): bool
     {
-        /* @var User */
+        /** @var User */
         $this->user = $token->getUser();
         $this->userId = $this->user->getId();
         $this->place = $place;