+++ tests/Controller/Rdv/RdvControllerTest.php
@@ -3,13 +3,14 @@
 namespace App\Tests\Controller\Rdv;
 
 use App\Entity\Support\Rdv;
+use App\Entity\Support\SupportGroup;
 use App\Tests\AppTestTrait;
-use Symfony\Component\DomCrawler\Crawler;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Bundle\FrameworkBundle\KernelBrowser;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Liip\TestFixturesBundle\Services\DatabaseToolCollection;
 use Liip\TestFixturesBundle\Services\DatabaseTools\AbstractDatabaseTool;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\DomCrawler\Crawler;
+use Symfony\Component\HttpFoundation\Response;
 
 class RdvControllerTest extends WebTestCase
 {
@@ -20,7 +21,7 @@ class RdvControllerTest extends WebTestCase
 
     /** @var AbstractDatabaseTool */
     protected $databaseTool;
-    
+
     /** @var array */
     protected $fixtures;
 
@@ -47,14 +48,13 @@ class RdvControllerTest extends WebTestCase
             dirname(__DIR__).'/../DataFixturesTest/RdvFixturesTest.yaml',
         ]);
 
-        $this->createLogin($this->fixtures['userRoleUser']);
-
         $this->supportGroup = $this->fixtures['supportGroup1'];
         $this->rdv = $this->fixtures['rdv1'];
     }
 
     public function testSearchRdvsIsSuccessful()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $this->client->request('GET', '/rdvs');
 
         // Page is up
@@ -74,6 +74,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testExportRdvsIsSuccessful()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $this->client->request('GET', '/rdvs');
 
         // Export with no result
@@ -93,6 +94,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testViewSupportListRdvsIsUp()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $id = $this->supportGroup->getId();
         $this->client->request('GET', "/support/$id/rdvs");
 
@@ -102,6 +104,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testCreateNewRdvIsSuccessful()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         /** @var Crawler */
         $crawler = $this->client->request('GET', '/calendar/month');
         $csrfToken = $crawler->filter('#rdv__token')->attr('value');
@@ -131,6 +134,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testCreateNewSupportRdvIsSuccessful()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $id = $this->supportGroup->getId();
         /** @var Crawler */
         $crawler = $this->client->request('GET', "/support/$id/calendar/month");
@@ -162,6 +166,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testGetRdv()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $id = $this->rdv->getId();
         $this->client->request('GET', "rdv/$id/get");
 
@@ -171,6 +176,7 @@ class RdvControllerTest extends WebTestCase
 
     public function testEditRdvIsSuccessful()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         /** @var Crawler */
         $crawler = $this->client->request('GET', '/calendar/month');
         $csrfToken = $crawler->filter('#rdv__token')->attr('value');
@@ -199,8 +205,30 @@ class RdvControllerTest extends WebTestCase
         $this->assertSame('RDV test edit', $content['rdv']['title']);
     }
 
+    public function testEditRdvWithOtherUserIsSuccessful()
+    {
+        $this->createLogin($this->fixtures['user4']);
+
+        $crawler = $this->client->request('GET', '/calendar/month');
+
+        $now = new \DateTime();
+        $this->client->request('POST', '/rdv/1/edit', [
+            'rdv' => [
+                'title' => 'RDV test edit',
+                'start' => $now->format('Y-m-d\TH:00'),
+                'end' => (clone $now)->modify('+1 hour')->format('Y-m-d\TH:00'),
+                '_token' => $crawler->filter('#rdv__token')->attr('value'),
+            ],
+        ]);
+
+        $this->assertResponseIsSuccessful();
+        $content = json_decode($this->client->getResponse()->getContent(), true);
+        $this->assertSame('RDV test edit', $content['rdv']['title']);
+    }
+
     public function testDeleteRdv()
     {
+        $this->createLogin($this->fixtures['userRoleUser']);
         $id = $this->rdv->getId();
         $this->client->request('GET', "/rdv/$id/delete");
 
@@ -211,7 +239,7 @@ class RdvControllerTest extends WebTestCase
     protected function tearDown(): void
     {
         parent::tearDown();
-        
+
         $this->client = null;
         $this->fixtures = null;
     }